tests/wpt/tests/layout-instability/visibility-hidden-layout-and-visible.html

<!DOCTYPE html>
<title>Layout Instability: visibility:hidden change with layout</title>
<link rel="help" href="https://wicg.github.io/layout-instability/" />
<div id="target" style="position: absolute; top: 0; width: 200px; height: 200px; visibility: hidden; background: blue"></div>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="resources/util.js"></script>
<script>

promise_test(async () => {
  const watcher = new ScoreWatcher;

  // Wait for the initial render to complete.
  await waitForAnimationFrames(2);

  // Shift target, for which no shift should be reported because it's hidden.
  target.style.top = '200px';
  target.style.visibility = 'visible';

  await waitForAnimationFrames(2);
  // No shift should be reported.
  assert_equals(watcher.score, 0);

  // Shift again, for which shift should be reported.
  target.style.top = '300px';

  await watcher.promise;
  const expectedScore = computeExpectedScore(200 * (200 + 100), 100);
  assert_equals(watcher.score, expectedScore);

}, 'visibility:hidden change with layout');

</script>
